feat: Add more options for helm chart - labels, annotations. securityContexts & cronjob arguments (#168)
Adding a few more options for the helm chart

- additional labels & annotations (includes making the use of labels more consistent)
- use the specified security context (includes some changes to the security contexts so they work)
- ability to pass extra args to the cron job

It should be possible to go a bit further with locking down the securityContexts in a later PR.  Specifically it should be possible to enable `runAsNonRoot`.  Right now that doesn't work because the `/app` directory permissions in the docker image mean you have to run as root.