fix: Make otel DB configurable (#559)
Even though we are supplying a full DSN (including the database) the CLICKHOUSE_DATABASE variable also needs to be defined.

I've set this to be `default` (by default) which corresponds to a CH database that has been created with no dedicated database (e.g. assets).

We might also look into parsing the DSN in Helm (and CloudFormation) to deduce the value for the CLICKHOUSE_DATABASE.