fix: Adding security context configuration (#587)
Following these general guidelines from [Trivy](https://avd.aquasec.com/misconfig/kubernetes/general/):

- Configuring pod security context:

```yaml
podSecurityContext:
  readOnlyRootFilesystem: false
  runAsNonRoot: true
  runAsUser: 3001
  runAsGroup: 3001
  fsGroup: 3001
  seccompProfile:
    type: RuntimeDefault
```

- Set some deployment resource:

```yaml
resources:
  limits:
    cpu: 500m
    memory: 512Mi
  requests:
    cpu: 500m
    memory: 512Mi
```

At this stage these are a guess, so we will have to tweak if needed. Once we've settled on some values we may want to comment them out as typically users are encouraged to consider their cluster needs and resource allocation. Our commented out values would serve as a baseline recommendation.

- NGINX configuration

To allow the various outputs from nginx to be routed to a writeable volume, I have included a copied and modified version of the configuration from the docker file. We mount this version in k8s. This allows us to target the volumes appropriately (and would also allow us to configure nginx values if needed). We could always absorb the necessary changes into the docker configuration and remove this, but for now it made sense to have it in the Helm deployment.

- Storage Volume

When enabling the `readOnlyRootFilesystem`, the plugin mirroring reported errors as it was not allowed to write to `/data/storage`. For now I've added an ephemeral volume to allow it to continue. In the next PR I'll add the option to make this a persistent volume.