feat: Adding pod disruption budget (#625)
The EKS cluster we are using is configured as an EKS auto cluster. This
results in nodes being auto-provisioned whenever pods exceed the
available resources. This happens when a sync is triggered, which can
lead to the platform and otel-collector pods being re-scheduled onto the
new node.

This PR adds Pod Disruption Budgets to ensure that both the platform and
otel-collector pods are not interrupted when node scaling occurs. This
will likely result in the original node being protected form eviction
whilst the sync jobs will be scheduled on a new, bigger node.


![image](https://github.com/user-attachments/assets/2a88abd6-be18-4c05-b95c-247cd92e98d4)